rework contacts import script to be less memory hungry
It seems we still hit the occasional segmentation fault on GitHub actions when
running `jq`, so I assume it's due to memory consumption (although it's hard to
tell and our files aren't big).

Several to work around this issue, but for now trying to use a simpler jq filter
that can use streaming mode (it didn't before, as I used slurp mode and not
really streaming), and then re-processing that with sed and awk to inject into
a payload template to send to sendinblue with curl.

This should work fine... *if* GitHub actions support awk...